FIX #734 Changed vectordb into vector_db
