+++ docs/advanced/query_configuration.mdx
@@ -8,7 +8,7 @@ title: '🔍 Query configurations'
 |-----------|-----------------------|---------------------------------|------------------------|
 | log_level | log level             | string                          | WARNING                |
 | embedding_fn| embedding function    | chromadb.utils.embedding_functions | \{text-embedding-ada-002\} |
-| db        | vector database (experimental) | BaseVectorDB               | ChromaDB               |
+| db        | vector database (experimental) | Basevector_db               | ChromaDB               |
 | collection_name | initial collection name for the database | string             | embedchain_store |
 | collect_metrics | collect anonymous telemetry data to improve embedchain | boolean | true |
