+++ embedchain/__init__.py
@@ -8,4 +8,4 @@ from embedchain.apps.Llama2App import Llama2App  # noqa: F401
 from embedchain.apps.open_source_app import OpenSourceApp  # noqa: F401
 from embedchain.apps.PersonApp import (PersonApp,  # noqa: F401
                                        PersonOpenSourceApp)
-from embedchain.vectordb.chroma import ChromaDB  # noqa: F401
+from embedchain.vector_db.chroma import ChromaDB  # noqa: F401
