+++ embedchain/config/apps/base_app_config.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from embedchain.config.base_config import BaseConfig
 from embedchain.helper.json_serializable import JSONSerializable
-from embedchain.vectordb.base import BaseVectorDB
+from embedchain.vector_db.base import Basevector_db
 
 
 class BaseAppConfig(BaseConfig, JSONSerializable):
@@ -14,7 +14,7 @@ class BaseAppConfig(BaseConfig, JSONSerializable):
     def __init__(
         self,
         log_level: str = "WARNING",
-        db: Optional[BaseVectorDB] = None,
+        db: Optional[Basevector_db] = None,
         id: Optional[str] = None,
         collect_metrics: bool = True,
         collection_name: Optional[str] = None,
@@ -27,7 +27,7 @@ class BaseAppConfig(BaseConfig, JSONSerializable):
         :type log_level: str, optional
         :param db: A database class. It is recommended to set this directly in the `App` class, not this config,
         defaults to None
-        :type db: Optional[BaseVectorDB], optional
+        :type db: Optional[Basevector_db], optional
         :param id: ID of the app. Document metadata will have this id., defaults to None
         :type id: Optional[str], optional
         :param collect_metrics: Send anonymous telemetry to improve embedchain, defaults to True
