+++ embedchain/config/apps/custom_app_config.py
@@ -3,7 +3,7 @@ from typing import Optional
 from dotenv import load_dotenv
 
 from embedchain.helper.json_serializable import register_deserializable
-from embedchain.vectordb.base import BaseVectorDB
+from embedchain.vector_db.base import Basevector_db
 
 from .base_app_config import BaseAppConfig
 
@@ -19,7 +19,7 @@ class CustomAppConfig(BaseAppConfig):
     def __init__(
         self,
         log_level: str = "WARNING",
-        db: Optional[BaseVectorDB] = None,
+        db: Optional[Basevector_db] = None,
         id: Optional[str] = None,
         collect_metrics: Optional[bool] = None,
         collection_name: Optional[str] = None,
@@ -32,7 +32,7 @@ class CustomAppConfig(BaseAppConfig):
         :type log_level: str, optional
         :param db: A database class. It is recommended to set this directly in the `CustomApp` class, not this config,
         defaults to None
-        :type db: Optional[BaseVectorDB], optional
+        :type db: Optional[Basevector_db], optional
         :param id: ID of the app. Document metadata will have this id., defaults to None
         :type id: Optional[str], optional
         :param collect_metrics: Send anonymous telemetry to improve embedchain, defaults to True
