+++ tests/apps/test_apps.py
@@ -5,8 +5,8 @@ from embedchain import App, CustomApp, Llama2App, OpenSourceApp
 from embedchain.config import ChromaDbConfig
 from embedchain.embedder.base import BaseEmbedder
 from embedchain.llm.base import BaseLlm
-from embedchain.vectordb.base import BaseVectorDB
-from embedchain.vectordb.chroma import ChromaDB
+from embedchain.vector_db.base import Basevector_db
+from embedchain.vector_db.chroma import ChromaDB
 
 
 class TestApps(unittest.TestCase):
@@ -18,26 +18,26 @@ class TestApps(unittest.TestCase):
     def test_app(self):
         app = App()
         self.assertIsInstance(app.llm, BaseLlm)
-        self.assertIsInstance(app.db, BaseVectorDB)
+        self.assertIsInstance(app.db, Basevector_db)
         self.assertIsInstance(app.embedder, BaseEmbedder)
 
     def test_custom_app(self):
         app = CustomApp()
         self.assertIsInstance(app.llm, BaseLlm)
-        self.assertIsInstance(app.db, BaseVectorDB)
+        self.assertIsInstance(app.db, Basevector_db)
         self.assertIsInstance(app.embedder, BaseEmbedder)
 
     def test_opensource_app(self):
         app = OpenSourceApp()
         self.assertIsInstance(app.llm, BaseLlm)
-        self.assertIsInstance(app.db, BaseVectorDB)
+        self.assertIsInstance(app.db, Basevector_db)
         self.assertIsInstance(app.embedder, BaseEmbedder)
 
     def test_llama2_app(self):
         os.environ["REPLICATE_API_TOKEN"] = "-"
         app = Llama2App()
         self.assertIsInstance(app.llm, BaseLlm)
-        self.assertIsInstance(app.db, BaseVectorDB)
+        self.assertIsInstance(app.db, Basevector_db)
         self.assertIsInstance(app.embedder, BaseEmbedder)
